model changes, added detail controller and view, fixed many rendering quirks and bugs

+++ services/ServiceEntities/UserData/HighChartResult.cs
@@ -9,7 +9,7 @@ namespace ServiceEntities.UserData
     public class HighChartResult
     {
         public string Name { get; set; }
-        public int Y { get; set; }
+        public decimal Y { get; set; }
         public string Color { get; set; }
     }
 }
